% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_process_req_file.R
\name{capture_repo_details}
\alias{capture_repo_details}
\title{Parse options relating to repositories}
\usage{
capture_repo_details(content)
}
\arguments{
\item{content}{a vector of repository option requirements}
}
\value{
vector of cran requirements
}
\description{
Parse options relating to repositories
}
