% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_process_req_file.R
\name{filter_repo_details}
\alias{filter_repo_details}
\title{Fetch the lines of a requirements file pertaining to repository options}
\usage{
filter_repo_details(content)
}
\arguments{
\item{content}{a vector of requirements}
}
\value{
vector of cran requirements
}
\description{
Fetch the lines of a requirements file pertaining to repository options
}
