% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_requirements.R
\name{generate_requirements}
\alias{generate_requirements}
\title{Generate a requirements.txt for your project}
\usage{
generate_requirements(glob_paths = "*.R", output_path = "requirements.txt",
  eq_sym = ">=", rm_missing = FALSE)
}
\arguments{
\item{glob_paths}{Character vector of patterns for relative or absolute filepaths. Missing values will be ignored.
See \code{?Sys.glob} for more details.}

\item{output_path}{String indicating where to write resulting requirements file to.}

\item{eq_sym}{The equality symbol to be used when writing requirements (i.e. package>=1.0.0).
Use \code{NULL} to not include package versions in your requirements file.}

\item{rm_missing}{Should packages not installed locally be exlcuded from output?}
}
\value{
Nothing is returned.  Results are written to \code{output_path}
}
\description{
Generate a requirements file from existing file(s) and installed packages
}
\examples{

\dontrun{
generate_requirements('R/*.R')
generate_requirements('R/*.R', output_path='my_requirements.txt')
generate_requirements('R/*.R', output_path='equal_to_requirements.txt', eq_sym='>=')
generate_requirements('R/*.R', output_path='versionless_requirements.txt', eq_sym=NULL)
generate_requirements('R/*.R', output_path='installed_requirements.txt', rm_missing=TRUE)
}
}
