% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rip_install_helpers.R
\name{get_archive_versions_extra}
\alias{get_archive_versions_extra}
\title{Get archived package versions from extra-CRAN sources}
\usage{
get_archive_versions_extra(package, repo)
}
\arguments{
\item{package}{name of package to fetch archived versions for}

\item{repo}{the url of the repo to look in}
}
\value{
character vector of archived package versions, NULL if none
}
\description{
Get archived package versions from extra-CRAN sources
}
