% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rip_install_helpers.R
\name{get_available_versions_extra}
\alias{get_available_versions_extra}
\title{Get packages from extra-CRAN sources}
\usage{
get_available_versions_extra(package, repo)
}
\arguments{
\item{package}{name of package to fetch versions for}

\item{repo}{the url of the repo to look in}
}
\value{
character vector of available package versions; NULL if package not found
}
\description{
Get packages from extra-CRAN sources
}
\examples{
get_available_versions_extra('emnTAW', 'http://lnxaws01.emn.com/EastmanCRAN/')
}
