% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rip_install_helpers.R
\name{get_current_versions_extra}
\alias{get_current_versions_extra}
\title{Get current versions of packages from extra-CRAN sources}
\usage{
get_current_versions_extra(repo)
}
\arguments{
\item{repo}{the url of the repo to look in}
}
\value{
data.frame of package names and versions, filled with NULL if unavailable
}
\description{
Get current versions of packages from extra-CRAN sources
}
