% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_requirements.R
\name{install_requirements}
\alias{install_requirements}
\title{Install project requirements}
\usage{
install_requirements(root_dir = getwd(), install = TRUE, packrat = FALSE,
  dryrun = FALSE, verbose = TRUE, repo = "https://cran.rstudio.com")
}
\arguments{
\item{root_dir}{Path to where requirements.txt is located}

\item{packrat}{Should project be initialized with packrat?}

\item{dryrun}{Flag if you just want to see what would happen if you ran
the function}

\item{verbose}{Flag to change verbosity}

\item{repo}{What repository should be used to install pacakges?}

\item{gen}{Should required packages be generated?}
}
\value{
invisible
}
\description{
Install (and optionally generate) required packages
}
\examples{
\dontrun{
requirements(dryrun = TRUE)
}
}
