% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requirements.R
\name{load_requirements}
\alias{load_requirements}
\title{Load project requirements}
\usage{
load_requirements(root_dir = getwd(), install = TRUE, packrat = FALSE,
  dryrun = FALSE, verbose = TRUE, repo = "https://cran.rstudio.com")
}
\arguments{
\item{root_dir}{Path to where requirements.txt is located}

\item{install}{Should required packages be installed?}

\item{packrat}{Should project be initialized with packrat?}

\item{dryrun}{Flag if you just want to see what would happen if you ran
the function}

\item{verbose}{Flag to change verbosity}

\item{repo}{What repository should be used to install pacakges?}
}
\value{
invisible
}
\description{
Load (and optionally install) required packages
}
\examples{
\dontrun{
requirements(dryrun = TRUE)
}
}
