% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_from_packrat.R
\name{packrat_to_requirements}
\alias{packrat_to_requirements}
\title{Generate a requirements file from a \code{packrat.lock} file}
\usage{
packrat_to_requirements(packrat_lock_path = "packrat/packrat.lock",
  output_path = "requirements.txt", eq_sym = ">=", append = TRUE)
}
\arguments{
\item{packrat_lock_path}{Path to \code{packrat.lock} to extract requirements from.}

\item{output_path}{String indicating where to write resulting requirements file to.}

\item{eq_sym}{The equality symbol to be used when writing requirements (i.e. package>=1.0.0).
Use \code{NULL} to not include package versions in your requirements file.}

\item{append}{Should requirements from \code{packrat_lock_path} be appended to end of \code{output_path}?
Overwrites contents of \code{output_path} if \code{FALSE}.}
}
\value{
Nothing is returned.  Results are written to \code{output_path}
}
\description{
Generate a requirements file from a \code{packrat.lock} file
}
\examples{

\dontrun{
packrat_to_requirements(output_path = "my_requirements.txt")
packrat_to_requirements(output_path = "versionless_requirements.txt", eq_sym = NULL)
packrat_to_requirements(output_path = "overwrite_requirements.txt", append = FALSE)
}
}
