% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_process_req_file.R
\name{parse_additional_repos}
\alias{parse_additional_repos}
\title{Parse a requirement file line to get the first URL after the option}
\usage{
parse_additional_repos(line)
}
\arguments{
\item{line}{a single line from a requirements file}
}
\value{
just the URL specifying the repository
}
\description{
Parse a requirement file line to get the first URL after the option
}
