% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_process_req_file.R
\name{process_find_links}
\alias{process_find_links}
\title{Read location specified by --find-links}
\usage{
process_find_links(loc)
}
\arguments{
\item{loc}{the location specified by --find-links}
}
\value{
the contents of the location
}
\description{
Read location specified by --find-links
}
\details{
Once a --find-link option has been parsed, a location is returned. This location can be a URL or a file. The
assumption here is that the location can be read as plain text, so a readLines is applied. If the location
cannot be read, then an error should be returned to the user as requirements may not be able to be satisified.
}
