% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.R
\name{req_dir}
\alias{req_dir}
\title{Extract requirements from a directory}
\usage{
req_dir(path = ".")
}
\arguments{
\item{path}{Path to recursively search}
}
\description{
Extracts requirements from all files in the directory and then ...
}
