% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{req_file}
\alias{req_file}
\title{Extract requirements from a file}
\usage{
req_file(path)
}
\arguments{
\item{path}{Path to file}
}
\description{
Extract requirements from a file
}
\examples{
path <- system.file("examples", "simple.R", package = "requirements")
req_file(path)
}
