% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\name{req_namespace}
\alias{req_namespace}
\title{Extract requirements from a `NAMESPACE` file}
\usage{
req_namespace(path)
}
\arguments{
\item{path}{Path to directory containing namespace file.}
}
\description{
Looks for `imports()` and `importFrom()` directives.
}
