% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requirements.R
\name{requirements}
\alias{requirements}
\title{Load project requirements}
\usage{
requirements(root_dir = getwd(), load = TRUE, packrat = FALSE,
  dryrun = FALSE, verbose = TRUE, repo = "https://cran.rstudio.com")
}
\arguments{
\item{root_dir}{Path to where requirements.txt is located}

\item{load}{Should the packages listed in requirements.txt be loaded to the search path?}

\item{packrat}{Should project be initialized with packrat?}

\item{dryrun}{Flag if you just want to see what would happen if your ran
the function}

\item{verbose}{Flag to change verbosity}

\item{repo}{What repository should be used to install pacakges?}
}
\value{
invisible
}
\description{
Load and (if necessary) install required packages
}
\examples{
\dontrun{
requirements(dryrun = TRUE)
}
}
