% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_requirements.R
\name{rip_install}
\alias{rip_install}
\title{Install project requirements}
\usage{
rip_install(requirements = "requirements.txt", packrat = FALSE,
  dryrun = FALSE, verbose = dryrun, repo = options()$repo[1],
  gen = !file.exists(requirements),
  glob_paths = file.path(dirname(requirements), "*.R"))
}
\arguments{
\item{requirements}{filename where requirements are stored}

\item{packrat}{Should project be initialized with packrat?}

\item{dryrun}{Flag if you just want to see what would happen if you ran
the function}

\item{verbose}{Flag to change verbosity}

\item{repo}{What repository should be used to install packages?}

\item{gen}{Should required packages be generated?}

\item{glob_paths}{Character vector of patterns for relative or absolute filepaths.
Missing values will be ignored. See ?Sys.glob for more details.}
}
\value{
invisible
}
\description{
Install (and optionally generate) required packages
}
\examples{
\dontrun{
requirements(dryrun = TRUE)
}
}
