% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\title{Install the packages provided to this function}
\usage{
install(packages = NULL, path_to_requirements = NULL)
}
\arguments{
\item{packages}{(Optional, String, also as vector)
The actual packages to install.}

\item{path_to_requirements}{(Optional, String) The path (including filename
and type ending) to a text file containing all the packages to install.}
}
\value{
(Boolean) `TRUE` if all packages are successfully installed,
 `FALSE` otherwise.
}
\description{
Install the packages provided to this function
}
