% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_requirements.R
\name{load_requirements}
\alias{load_requirements}
\title{Load project requirements}
\usage{
load_requirements(requirements = file.path(getwd(), "requirements.txt"),
  install = TRUE, packrat = FALSE, dryrun = FALSE, verbose = TRUE,
  repo = "https://cran.rstudio.com")
}
\arguments{
\item{requirements}{filename where requirements are stored}

\item{install}{Should required packages be installed?}

\item{packrat}{Should project be initialized with packrat?}

\item{dryrun}{Flag if you just want to see what would happen if you ran
the function}

\item{verbose}{Flag to change verbosity}

\item{repo}{What repository should be used to install pacakges?}
}
\value{
invisible
}
\description{
Load (and optionally install) required packages
}
\examples{
\dontrun{
requirements(dryrun = TRUE)
}
}
