% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{req_code}
\alias{req_code}
\title{Find requirements from a block of code}
\usage{
req_code(x)
}
\arguments{
\item{x}{Code to examine. Supports unquoting.}
}
\description{
Looks for `::`, `:::`, `library()`, `require()`, `requireNamespace()`,
and `loadNamespace()`.
}
\examples{
req_code(library("ralng"))
}
