% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{req_file}
\alias{req_file}
\title{Extract requirements from a file}
\usage{
req_file(path)
}
\arguments{
\item{path}{Path to file}
}
\description{
* `.R`: extracts requirements from parsed code

* `.Rmd` & `.Rpres`: requirements from chunks (parsed using regular
  expressions to avoid dependency on knitr package). If rmarkdown package is
  installed, will also add requirements from custom output type.

* `.Rnw`: tangles the document and then extracts from `.R` file.
}
\examples{
path_r <- system.file("examples", "simple.R", package = "requirements")
path_rmd <- system.file("examples", "simple.Rmd", package = "requirements")

req_file(path_r)
req_file(path_rmd)
}
