% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/necessity-package.R
\docType{package}
\name{requirements-package}
\alias{requirements}
\alias{requirements-package}
\title{requirements: Find Packages Required for Code to Run}
\description{
Figure out what packages are required to run a given R expression, 
    file ('.R', '.Rmd', '.Rnw', ...), or a directory of files. Includes 
    heuristics for common implicit dependencies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hadley/requirements}
  \item Report bugs at \url{https://github.com/hadley/requirements/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
