% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rip_install.R
\name{rip_install}
\alias{rip_install}
\title{Install project requirements}
\usage{
rip_install(
  requirements = "requirements.txt",
  packrat = FALSE,
  dryrun = FALSE,
  verbose = dryrun,
  repo = options()$repo[1],
  gen = !file.exists(requirements),
  inspect_files = c("R", "Rmd", "Rpres", "lock"),
  path = dirname(requirements),
  ...
)
}
\arguments{
\item{requirements}{filename where requirements are stored}

\item{packrat}{Should project be initialized with packrat?}

\item{dryrun}{Flag if you just want to see what would happen if you ran
the function}

\item{verbose}{Flag to change verbosity}

\item{repo}{What repository should be used to install packages?}

\item{gen}{Should required packages be generated?}

\item{inspect_files}{Character vector of file extensions to search for dependencies in.
Officially supports the following extensions: `c("R", "Rmd", "Rpres", "lock")`.
Other extensions will be processed as if they were `.R` files.}

\item{path}{Location where extensions from `inspect_files` should be searched for.
(see `path` from `?list.files`).}

\item{...}{Additional arguments to be passed to `rip_freeze()`}
}
\value{
invisible
}
\description{
Install (and optionally generate) required packages
}
\examples{
\dontrun{
requirements(dryrun = TRUE)
}
}
